Fixes, trying to follow better practices

+++ day1/main.cpp
@@ -19,15 +19,15 @@ int main() {
   sort(v2.begin(), v2.end());
 
   int totalDistance = 0;
-  for (int i = 0; i < v1.size(); i++) {
+  for (size_t i = 0; i < v1.size(); i++) {
     totalDistance += abs(v1[i] - v2[i]);
   }
   std::cout << "Part One:\nTotal Distance: " << totalDistance << std::endl;
 
   int similarityScore = 0;
-  for (int i = 0; i < v1.size(); i++) {
+  for (size_t i = 0; i < v1.size(); i++) {
     int dups = 0;
-    for (int j = 0; j < v1.size(); j++) {
+    for (size_t j = 0; j < v1.size(); j++) {
       if (v1[i] == v2[j]) {
         dups++;
       }
